backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.8.2
pydantic-settings
python-dotenv==1.0.1
PyYAML==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
